src/preprocessor/requirements.txt
google-cloud-pubsub==2.23.0
google-genai==1.10.0
weaviate-client==4.9.0
